requirements.txt
# Core dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
python-dotenv>=0.19.0
PyYAML>=5.4.1

# Database
SQLAlchemy>=1.4.23
alembic>=1.7.1
psycopg2-binary>=2.9.1

# AI and ML
torch>=1.9.0
transformers>=4.9.2
numpy>=1.21.2
scipy>=1.7.1

# Blockchain
web3>=5.24.0
eth-account>=0.5.5
eth-typing>=2.2.2

# Security
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=3.2.0
python-multipart>=0.0.5

# Caching and messaging
redis>=3.5.3
aioredis>=2.0.0
celery>=5.1.2

# Storage
boto3>=1.18.21
aiofiles>=0.7.0

# Monitoring and logging
prometheus-client>=0.11.0
statsd>=3.3.0
python-json-logger>=2.0.2

# Testing
pytest>=6.2.5
pytest-asyncio>=0.15.1
pytest-cov>=2.12.1
httpx>=0.18.2

# Development
black>=21.7b0
isort>=5.9.3
flake8>=3.9.2
mypy>=0.910
pre-commit>=2.15.0

# Documentation
Sphinx>=4.1.2
sphinx-rtd-theme>=0.5.2
mkdocs>=1.2.2
mkdocs-material>=7.2.6 